move TrajectoryStopReasons to DataFormats/TrackCandidate in order to break circular dependence
