+++ TrackingTools/PatternTools/BuildFile.xml
@@ -3,6 +3,7 @@
 <use name="DataFormats/GeometrySurface"/>
 <use name="DataFormats/GeometryVector"/>
 <use name="DataFormats/TrackReco"/>
+<use name="DataFormats/TrackCandidate"/>
 <use name="DataFormats/TrackerRecHit2D"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/Utilities"/>
